remove indexer, move minhash_lsh (still disabled) to blocking
